Revert "Restructure project to separate core provider APIs from non-provider …"
This reverts commit f1154b06323359ba8915db6fdebbc172af0c6271.
