--- src/builtins/native/mod.rs
@@ -1,32 +1,0 @@
-//! This module implements native Rust wrappers for the Temporal builtins.
-
-pub(crate) mod timezone;
-
-mod date;
-mod datetime;
-mod duration;
-mod instant;
-mod now;
-pub(crate) mod options;
-mod time;
-mod zoneddatetime;
-
-#[doc(inline)]
-pub use date::PlainDate;
-#[doc(inline)]
-pub use datetime::PlainDateTime;
-#[doc(inline)]
-pub use duration::Duration;
-#[doc(inline)]
-pub use instant::Instant;
-#[doc(inline)]
-pub use now::Now;
-#[doc(inline)]
-pub use time::PlainTime;
-#[doc(inline)]
-pub use zoneddatetime::ZonedDateTime;
-
-pub use crate::builtins::core::{
-    calendar, DateDuration, PartialDate, PartialDateTime, PartialTime, PartialZonedDateTime,
-    PlainMonthDay, PlainYearMonth, TimeDuration,
-};
