--- src/builtins/native/zoneddatetime.rs
@@ -1,445 +1,0 @@
-use super::timezone::TZ_PROVIDER;
-use crate::{
-    builtins::core as temporal_core,
-    options::{
-        ArithmeticOverflow, DifferenceSettings, Disambiguation, DisplayCalendar, DisplayOffset,
-        DisplayTimeZone, OffsetDisambiguation, ToStringRoundingOptions,
-    },
-    Calendar, Duration, PlainDate, PlainDateTime, PlainTime, TemporalError, TemporalResult,
-    TimeZone,
-};
-use alloc::string::String;
-use tinystr::TinyAsciiStr;
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ZonedDateTime(pub(crate) temporal_core::ZonedDateTime);
-
-impl core::fmt::Display for ZonedDateTime {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.write_str(
-            &self
-                .to_ixdtf_string(
-                    DisplayOffset::Auto,
-                    DisplayTimeZone::Auto,
-                    DisplayCalendar::Auto,
-                    ToStringRoundingOptions::default(),
-                )
-                .expect("A valid ZonedDateTime string with default options."),
-        )
-    }
-}
-
-impl From<temporal_core::ZonedDateTime> for ZonedDateTime {
-    fn from(value: temporal_core::ZonedDateTime) -> Self {
-        Self(value)
-    }
-}
-
-impl ZonedDateTime {
-    #[inline]
-    pub fn try_new(nanos: i128, calendar: Calendar, tz: TimeZone) -> TemporalResult<Self> {
-        temporal_core::ZonedDateTime::try_new(nanos, calendar, tz).map(Into::into)
-    }
-
-    pub fn calendar(&self) -> &Calendar {
-        self.0.calendar()
-    }
-
-    pub fn timezone(&self) -> &TimeZone {
-        self.0.timezone()
-    }
-}
-
-// ===== Experimental TZ_PROVIDER accessor implementations =====
-
-impl ZonedDateTime {
-    pub fn year(&self) -> TemporalResult<i32> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.year_with_provider(&*provider)
-    }
-
-    pub fn month(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.month_with_provider(&*provider)
-    }
-
-    pub fn month_code(&self) -> TemporalResult<TinyAsciiStr<4>> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.month_code_with_provider(&*provider)
-    }
-
-    pub fn day(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.day_with_provider(&*provider)
-    }
-
-    pub fn hour(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.hour_with_provider(&*provider)
-    }
-
-    pub fn minute(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.minute_with_provider(&*provider)
-    }
-
-    pub fn second(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.second_with_provider(&*provider)
-    }
-
-    pub fn millisecond(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.millisecond_with_provider(&*provider)
-    }
-
-    pub fn microsecond(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.millisecond_with_provider(&*provider)
-    }
-
-    pub fn nanosecond(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-
-        self.0.millisecond_with_provider(&*provider)
-    }
-}
-
-// ==== Experimental TZ_PROVIDER calendar method implementations ====
-
-impl ZonedDateTime {
-    pub fn era(&self) -> TemporalResult<Option<TinyAsciiStr<16>>> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.era_with_provider(&*provider)
-    }
-
-    pub fn era_year(&self) -> TemporalResult<Option<i32>> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.era_year_with_provider(&*provider)
-    }
-
-    /// Returns the calendar day of week value.
-    pub fn day_of_week(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.day_of_week_with_provider(&*provider)
-    }
-
-    /// Returns the calendar day of year value.
-    pub fn day_of_year(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.day_of_year_with_provider(&*provider)
-    }
-
-    /// Returns the calendar week of year value.
-    pub fn week_of_year(&self) -> TemporalResult<Option<u16>> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.week_of_year_with_provider(&*provider)
-    }
-
-    /// Returns the calendar year of week value.
-    pub fn year_of_week(&self) -> TemporalResult<Option<i32>> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.year_of_week_with_provider(&*provider)
-    }
-
-    /// Returns the calendar days in week value.
-    pub fn days_in_week(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.days_in_week_with_provider(&*provider)
-    }
-
-    /// Returns the calendar days in month value.
-    pub fn days_in_month(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.days_in_month_with_provider(&*provider)
-    }
-
-    /// Returns the calendar days in year value.
-    pub fn days_in_year(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.days_in_year_with_provider(&*provider)
-    }
-
-    /// Returns the calendar months in year value.
-    pub fn months_in_year(&self) -> TemporalResult<u16> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.months_in_year_with_provider(&*provider)
-    }
-
-    /// Returns returns whether the date in a leap year for the given calendar.
-    pub fn in_leap_year(&self) -> TemporalResult<bool> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.in_leap_year_with_provider(&*provider)
-    }
-
-    pub fn hours_in_day(&self) -> TemporalResult<u8> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.hours_in_day_with_provider(&*provider)
-    }
-}
-
-// ==== Experimental TZ_PROVIDER method implementations ====
-
-impl ZonedDateTime {
-    /// Creates a new `ZonedDateTime` from the current `ZonedDateTime`
-    /// combined with the provided `TimeZone`.
-    pub fn with_plain_time(&self, time: PlainTime) -> TemporalResult<Self> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .with_plain_time_and_provider(time.0, &*provider)
-            .map(Into::into)
-    }
-
-    pub fn add(
-        &self,
-        duration: &Duration,
-        overflow: Option<ArithmeticOverflow>,
-    ) -> TemporalResult<Self> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .add_with_provider(&duration.0, overflow, &*provider)
-            .map(Into::into)
-    }
-
-    pub fn subtract(
-        &self,
-        duration: &Duration,
-        overflow: Option<ArithmeticOverflow>,
-    ) -> TemporalResult<Self> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .subtract_with_provider(&duration.0, overflow, &*provider)
-            .map(Into::into)
-    }
-
-    /// Returns a [`Duration`] representing the period of time from this `ZonedDateTime` since the other `ZonedDateTime`.
-    pub fn since(&self, other: &Self, options: DifferenceSettings) -> TemporalResult<Duration> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .since_with_provider(&other.0, options, &*provider)
-            .map(Into::into)
-    }
-
-    /// Returns a [`Duration`] representing the period of time from this `ZonedDateTime` since the other `ZonedDateTime`.
-    pub fn until(&self, other: &Self, options: DifferenceSettings) -> TemporalResult<Duration> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .until_with_provider(&other.0, options, &*provider)
-            .map(Into::into)
-    }
-
-    pub fn start_of_day(&self) -> TemporalResult<Self> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .start_of_day_with_provider(&*provider)
-            .map(Into::into)
-    }
-
-    /// Creates a new [`PlainDate`] from this `ZonedDateTime`.
-    pub fn to_plain_date(&self) -> TemporalResult<PlainDate> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .to_plain_date_with_provider(&*provider)
-            .map(Into::into)
-    }
-
-    /// Creates a new [`PlainTime`] from this `ZonedDateTime`.
-    pub fn to_plain_time(&self) -> TemporalResult<PlainTime> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .to_plain_time_with_provider(&*provider)
-            .map(Into::into)
-    }
-
-    /// Creates a new [`PlainDateTime`] from this `ZonedDateTime`.
-    pub fn to_plain_datetime(&self) -> TemporalResult<PlainDateTime> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0
-            .to_plain_datetime_with_provider(&*provider)
-            .map(Into::into)
-    }
-
-    /// Returns a RFC9557 (IXDTF) string with the provided options.
-    pub fn to_ixdtf_string(
-        &self,
-        display_offset: DisplayOffset,
-        display_timezone: DisplayTimeZone,
-        display_calendar: DisplayCalendar,
-        options: ToStringRoundingOptions,
-    ) -> TemporalResult<String> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        self.0.to_ixdtf_string_with_provider(
-            display_offset,
-            display_timezone,
-            display_calendar,
-            options,
-            &*provider,
-        )
-    }
-
-    pub fn from_str(
-        source: &str,
-        disambiguation: Disambiguation,
-        offset_option: OffsetDisambiguation,
-    ) -> TemporalResult<Self> {
-        let provider = TZ_PROVIDER
-            .lock()
-            .map_err(|_| TemporalError::general("Unable to acquire lock"))?;
-        temporal_core::ZonedDateTime::from_str_with_provider(
-            source,
-            disambiguation,
-            offset_option,
-            &*provider,
-        )
-        .map(Into::into)
-    }
-}
-
-mod tests {
-    #[cfg(not(target_os = "windows"))]
-    #[test]
-    fn static_tzdb_zdt_test() {
-        use super::ZonedDateTime;
-        use crate::{Calendar, TimeZone};
-        use core::str::FromStr;
-
-        let nov_30_2023_utc = 1_701_308_952_000_000_000i128;
-
-        let zdt = ZonedDateTime::try_new(
-            nov_30_2023_utc,
-            Calendar::from_str("iso8601").unwrap(),
-            TimeZone::try_from_str("Z").unwrap(),
-        )
-        .unwrap();
-
-        assert_eq!(zdt.year().unwrap(), 2023);
-        assert_eq!(zdt.month().unwrap(), 11);
-        assert_eq!(zdt.day().unwrap(), 30);
-        assert_eq!(zdt.hour().unwrap(), 1);
-        assert_eq!(zdt.minute().unwrap(), 49);
-        assert_eq!(zdt.second().unwrap(), 12);
-
-        let zdt_minus_five = ZonedDateTime::try_new(
-            nov_30_2023_utc,
-            Calendar::from_str("iso8601").unwrap(),
-            TimeZone::try_from_str("America/New_York").unwrap(),
-        )
-        .unwrap();
-
-        assert_eq!(zdt_minus_five.year().unwrap(), 2023);
-        assert_eq!(zdt_minus_five.month().unwrap(), 11);
-        assert_eq!(zdt_minus_five.day().unwrap(), 29);
-        assert_eq!(zdt_minus_five.hour().unwrap(), 20);
-        assert_eq!(zdt_minus_five.minute().unwrap(), 49);
-        assert_eq!(zdt_minus_five.second().unwrap(), 12);
-
-        let zdt_plus_eleven = ZonedDateTime::try_new(
-            nov_30_2023_utc,
-            Calendar::from_str("iso8601").unwrap(),
-            TimeZone::try_from_str("Australia/Sydney").unwrap(),
-        )
-        .unwrap();
-
-        assert_eq!(zdt_plus_eleven.year().unwrap(), 2023);
-        assert_eq!(zdt_plus_eleven.month().unwrap(), 11);
-        assert_eq!(zdt_plus_eleven.day().unwrap(), 30);
-        assert_eq!(zdt_plus_eleven.hour().unwrap(), 12);
-        assert_eq!(zdt_plus_eleven.minute().unwrap(), 49);
-        assert_eq!(zdt_plus_eleven.second().unwrap(), 12);
-    }
-
-    #[cfg(not(target_os = "windows"))]
-    #[test]
-    fn basic_zdt_add() {
-        use super::ZonedDateTime;
-        use crate::{Calendar, Duration, TimeZone};
-
-        let zdt =
-            ZonedDateTime::try_new(-560174321098766, Calendar::default(), TimeZone::default())
-                .unwrap();
-        let d = Duration::new(
-            0.into(),
-            0.into(),
-            0.into(),
-            0.into(),
-            240.into(),
-            0.into(),
-            0.into(),
-            0.into(),
-            0.into(),
-            800.into(),
-        )
-        .unwrap();
-        // "1970-01-04T12:23:45.678902034+00:00[UTC]"
-        let expected =
-            ZonedDateTime::try_new(303825678902034, Calendar::default(), TimeZone::default())
-                .unwrap();
-
-        let result = zdt.add(&d, None).unwrap();
-        assert_eq!(result, expected);
-    }
-}
