+++ src/components/date.rs
@@ -1,9 +1,7 @@
 //! This module implements `Date` and any directly related algorithms.
 
 use crate::{
-    builtins::core::{
-        calendar::Calendar, duration::DateDuration, Duration, PlainDateTime, PlainTime,
-    },
+    components::{calendar::Calendar, duration::DateDuration, Duration, PlainDateTime},
     iso::{IsoDate, IsoDateTime, IsoTime},
     options::{
         ArithmeticOverflow, DifferenceOperation, DifferenceSettings, DisplayCalendar,
@@ -11,7 +9,6 @@ use crate::{
     },
     parsers::{parse_date_time, IxdtfStringBuilder},
     primitive::FiniteF64,
-    provider::NeverProvider,
     TemporalError, TemporalResult, TemporalUnwrap, TimeZone,
 };
 use alloc::{format, string::String};
@@ -20,7 +17,8 @@ use core::str::FromStr;
 use super::{
     calendar::{ascii_four_to_integer, month_to_month_code},
     duration::{normalized::NormalizedDurationRecord, TimeDuration},
-    PlainMonthDay, PlainYearMonth,
+    timezone::NeverProvider,
+    PlainMonthDay, PlainTime, PlainYearMonth,
 };
 use tinystr::TinyAsciiStr;
