+++ src/components/duration/normalized.rs
@@ -5,14 +5,16 @@ use core::{num::NonZeroU128, ops::Add};
 use num_traits::{AsPrimitive, Euclid, FromPrimitive};
 
 use crate::{
-    builtins::core::{timezone::TimeZone, PlainDate, PlainDateTime},
+    components::{
+        timezone::{TimeZone, TimeZoneProvider},
+        PlainDate, PlainDateTime,
+    },
     iso::{IsoDate, IsoDateTime},
     options::{
         ArithmeticOverflow, Disambiguation, ResolvedRoundingOptions, TemporalRoundingMode,
         TemporalUnit,
     },
     primitive::FiniteF64,
-    provider::TimeZoneProvider,
     rounding::{IncrementRounder, Round},
     TemporalError, TemporalResult, TemporalUnwrap, NS_PER_DAY,
 };
