+++ src/tzdb.rs
@@ -47,10 +47,10 @@ use tzif::{
     },
 };
 
+use crate::components::timezone::TimeZoneOffset;
 use crate::{
+    components::{timezone::TimeZoneProvider, EpochNanoseconds},
     iso::IsoDateTime,
-    provider::{TimeZoneOffset, TimeZoneProvider},
-    time::EpochNanoseconds,
     utils, TemporalError, TemporalResult,
 };
